refactor: add typing + docstrings
