--- src/js/users/components/Permission.jsx
@@ -1,31 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-import { BoxGroupSection, Icon } from "../../base";
-
-const StyledPermissionIcon = styled(Icon)`
-    margin-right: 15px;
-`;
-
-const PermissionIcon = ({ value }) => (
-    <StyledPermissionIcon name={value ? "check" : "times"} color={value ? "green" : "red"} fixedWidth />
-);
-
-const StyledPermissionItem = styled(BoxGroupSection)`
-    align-items: center;
-    display: flex;
-`;
-
-export const PermissionItem = ({ permission, value }) => {
-    if (value) {
-        return (
-            <StyledPermissionItem aria-label={`${permission}:${value}`}>
-                <PermissionIcon value={value} /> <code>{permission}</code>
-            </StyledPermissionItem>
-        );
-    }
-    return (
-        <StyledPermissionItem aria-label={`${permission}:${value}`}>
-            <PermissionIcon value={value} /> <code>{permission}</code>
-        </StyledPermissionItem>
-    );
-};
